Add an inheritance relationship for entities (#17)
