+++ src/main/java/org/gridsuite/modification/server/dto/ModificationInfos.java
@@ -1,8 +1,8 @@
-/**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+/*
+  Copyright (c) 2021, RTE (http://www.rte-france.com)
+  This Source Code Form is subject to the terms of the Mozilla Public
+  License, v. 2.0. If a copy of the MPL was not distributed with this
+  file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 package org.gridsuite.modification.server.dto;
 
@@ -12,6 +12,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
+import org.gridsuite.modification.server.ModificationType;
 
 import java.time.ZonedDateTime;
 import java.util.UUID;
@@ -30,4 +31,7 @@ public class ModificationInfos {
 
     @ApiModelProperty("Modification date")
     ZonedDateTime date;
+
+    @ApiModelProperty("Modification type")
+    ModificationType type;
 }
