+++ src/main/java/org/gridsuite/modification/server/entities/AbstractModificationEntity.java
@@ -1,0 +1,35 @@
+/*
+  Copyright (c) 2021, RTE (http://www.rte-france.com)
+  This Source Code Form is subject to the terms of the Mozilla Public
+  License, v. 2.0. If a copy of the MPL was not distributed with this
+  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.modification.server.entities;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.ZonedDateTime;
+import java.util.UUID;
+
+/**
+ * @author Slimane Amar <slimane.amar at rte-france.com>
+ */
+@NoArgsConstructor
+@Getter
+@Entity
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+@Table(name = "modification")
+public abstract class AbstractModificationEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    protected UUID id;
+
+    @Column(name = "date")
+    protected ZonedDateTime date;
+
+    @Column(name = "type")
+    protected String type;
+}
