+++ src/main/java/org/gridsuite/modification/server/entities/ElementaryModificationEntity.java
@@ -1,19 +1,21 @@
-/**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+/*
+  Copyright (c) 2021, RTE (http://www.rte-france.com)
+  This Source Code Form is subject to the terms of the Mozilla Public
+  License, v. 2.0. If a copy of the MPL was not distributed with this
+  file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 package org.gridsuite.modification.server.entities;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.gridsuite.modification.server.ModificationType;
 import org.gridsuite.modification.server.dto.ElementaryModificationInfos;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
-import java.util.UUID;
 
 /**
  * @author Slimane Amar <slimane.amar at rte-france.com>
@@ -22,15 +24,7 @@ import java.util.UUID;
 @Getter
 @Entity
 @Table(name = "elementaryModification")
-public class ElementaryModificationEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    private UUID id;
-
-    @Column(name = "date")
-    private ZonedDateTime date;
+public class ElementaryModificationEntity extends AbstractModificationEntity {
 
     @Column(name = "equipmentId")
     private String equipmentId;
@@ -46,6 +40,7 @@ public class ElementaryModificationEntity {
 
     public ElementaryModificationEntity(String equipmentId, String equipmentName, String equipmentAttributeName, String equipmentAttributeValue) {
         this.date = ZonedDateTime.now(ZoneOffset.UTC);
+        this.type = ModificationType.ELEMENTARY.name();
         this.equipmentId = equipmentId;
         this.equipmentName = equipmentName;
         this.equipmentAttributeName = equipmentAttributeName;
@@ -56,6 +51,7 @@ public class ElementaryModificationEntity {
         return ElementaryModificationInfos.builder()
                 .id(this.id)
                 .date(this.date)
+                .type(ModificationType.valueOf(this.type))
                 .equipmentId(this.equipmentId)
                 .equipmentName(this.equipmentName)
                 .equipmentAttributeName(this.equipmentAttributeName)
