--- src/test/resources/database.properties
@@ -1,4 +1,0 @@
-driverClassName=org.h2.Driver
-url=jdbc:h2:mem:db;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false
-username=postgres
-password=postgres
